Moved response_io utilities to a seperate module.
